requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.34.0
websockets==14.1
pydantic>=2.0.0
python-dotenv>=1.0.0
httpx>=0.27.0

# AI Models
anthropic>=0.45.0
google-generativeai>=0.8.0
openai>=1.50.0

# Text-to-Speech
elevenlabs>=1.0.0

# WhatsApp via Twilio
twilio>=9.0.0

# Audio processing
numpy>=1.26.0
